% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{addProgressPanel}
\alias{addProgressPanel}
\title{Adds elements to a Progress Panel}
\usage{
addProgressPanel(session = getDefaultReactiveDomain(), inputId, id, image)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{id}{The ids of the added elements.}

\item{image}{Paths of the added images.}
}
\description{
Adds elements to a Progress Panel by given id and path to image
}
\examples{
\code{
addProgressPanel(inputId="progressPanel", id=10, image="images/img10.jpg")
}

}
