% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayasGroupedButtons.R
\name{bayasGroupedButtons}
\alias{bayasGroupedButtons}
\title{Grouped Action Buttons Inputs}
\usage{
\code{bayasGroupedButtons(
  inputId,
  btnNames,
  btnValues,
  selected=NULL,
  columns=3,
  btnStyle="",
  outerStyle=""
)}
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{btnNames}{A string vector for button labels. Have to be unique names.}

\item{btnValues}{A vector of same length as btnNames containing the values
that should returned when clicking on the buttons.}

\item{selected}{The selected selected button. NULL for no selected selection.}

\item{columns}{The number of elements per row.}

\item{btnStyle}{Optional styling for the buttons.}

\item{outerStyle}{Optional styling for the outer div.}
}
\description{
Create a group of actions buttons with an optional default selected button.
Works similar to grouped radio buttons.
}
\section{Server return}{

Using the common shiny access via 'input'.
Returns the corresponding value given by btnValues.
}

\examples{
\code{
bayasGroupedButtons("groupedBtn", c("small","medium","big"), selected="small",
 columns = 2)
}
}
