% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayasSidePanel.R
\name{bayasSidePanel}
\alias{bayasSidePanel}
\title{Layout a sidebar and main area}
\usage{
bayasSidePanel(
  inputId,
  sidePanel,
  mainPanel,
  sidePanelClass = NULL,
  mainPanelClass = NULL,
  sidePanelStyle = NULL,
  mainPanelStyle = NULL
)
}
\arguments{
\item{inputId}{The id of the input object. Used to open/close from server side.}

\item{sidePanel}{A div containing the left sidebar panel.}

\item{mainPanel}{A div containing the main panel.}

\item{sidePanelClass}{Additional styling of the sidebar panel}

\item{mainPanelClass}{Additional styling of the main panel}

\item{sidePanelStyle}{Additional styling of the sidebar panel}

\item{mainPanelStyle}{Additional styling of the main panel}
}
\description{
A sidebar layout similar to shinys 'sidebarLayout' but with a collapsable sidebar panel
}
\note{
see: \link[shinybayas]{updateBayasSidePanel}
}
