% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{hideProgressPanel}
\alias{hideProgressPanel}
\title{Hides an existing Progress Panel}
\usage{
hideProgressPanel(session = getDefaultReactiveDomain(), inputId)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}
}
\description{
Hides an existing Progress Panel
}
\examples{
\code{
hideProgressPanel(inputId="groupedBtn")
}

}
