% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{progressPanel}
\alias{progressPanel}
\title{Progress Panel}
\usage{
\code{progressPanel(
  inputId,
  initIds = NULL,
  initImages = NULL,
  visible = F,
  title = NULL,
  footer = progressPanelButton()
)}
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{initIds}{Ids for the optional initial elements of the overall panel.}

\item{initImages}{Paths of images to the optional initial elements of the progress panel.}

\item{visible}{Should be the panel visible at start?}

\item{title}{The title, if NULL for no title.}

\item{footer}{The footer, if NULL for no footer.}
}
\description{
Creates a progress panel that behaves like a modal.
Add removable elements that shows the current progress.
}
\section{Server return}{

Using the common shiny access via 'input'.
Returns the user given id of the last removed element.
}

\examples{
\code{
progressPanel("progressPanel", initIds=c(1,2), initImages=c("folder/img1.jpg","folder/img2.jpg",
visible = F, title = "Added elements"),
}
}
