% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{progressPanelButton}
\alias{progressPanelButton}
\title{Creates a close button for a Progress Panel}
\usage{
progressPanelButton()
}
\description{
Creates a close button for a Progress Panel
}
\examples{
\code{
progressPanelButton()
}

}
