% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{removeProgressPanel}
\alias{removeProgressPanel}
\title{Removes elements of a Progress Panel}
\usage{
removeProgressPanel(session = getDefaultReactiveDomain(), inputId, ids)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{id}{The ids of the elements that should be removed.}
}
\description{
Removes elements of a Progress Panel by given id
}
\examples{
\code{
removeProgressPanel(inputId="progressPanel", id=10)
}
}
