% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayasGroupedButtons.R
\name{updateBayasGroupedButtons}
\alias{updateBayasGroupedButtons}
\title{Update Grouped Action Buttons Inputs}
\usage{
updateBayasGroupedButtons(
  session = getDefaultReactiveDomain(),
  inputId,
  selected,
  value = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.
Default is getDefaultReactiveDomain().}

\item{inputId}{The id of the input object.}

\item{selected}{The button that should be selected.}
}
\description{
Updates an existing grouped button created with \code{groupedButtons()}
}
\examples{
\code{
updateBayasGroupedButtons(inputId="groupedBtn", selected="big")
}
}
