% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayasSidePanel.R
\name{updateBayasSidePanel}
\alias{updateBayasSidePanel}
\title{Updates a 'bayasSidePanel'}
\usage{
updateBayasSidePanel(session, inputId, status = c("open", "close"))
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{status}{c("open","close")}
}
\description{
Updates the status (open, close) of a 'bayasSidePanel'.
}
\note{
see: \link[shinybayas]{bayasSidePanel}
}
\examples{
\code{
updateBayasSidePanel(session = getDefaultReactiveDomain(), inputId = "bayasSidebarPanel",
status = "close")
}

}
