% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayasTextInput.R
\name{updateBayasTextInput}
\alias{updateBayasTextInput}
\title{Change the value of a text input on the client}
\usage{
\code{
updateTextInput(
session = getDefaultReactiveDomain(),
inputId,
label = NULL,
value = NULL,
placeholder = NULL
)}
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control. Internet Explorer 8 and 9 do not support this option.}

\item{width}{The width of the input, e.g. '400px', or '100%'}
}
\description{
Change the value of a text input on the client
}
