% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressPanel.R
\name{updateProgressPanel}
\alias{updateProgressPanel}
\title{Updates Progress Panel}
\usage{
updateProgressPanel(session = getDefaultReactiveDomain(), inputId, ids, images)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{id}{The ids of the changed elements.}

\item{image}{Paths of the changed images.}
}
\description{
Updates an existing Progress Panel
}
\examples{
\code{
updateProgressPanel(inputId="groupedBtn", id=10, image="images/img10.jpg")
}

}
